refactor(core): share client logic
