+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 - Bump project version to `0.1.4`.
+- Extracted common logic from `Client` and `AsyncClient` into new `HTTPClientBase`.
 
 ## [0.1.4]
