+++ imednet/core/client.py
@@ -13,18 +13,16 @@ from __future__ import annotations
 
 import logging
 from types import TracebackType
-from typing import Any, Callable, Dict, Optional, Union, cast
+from typing import Any, Dict, Optional, cast
 
 import httpx
-from tenacity import RetryCallState
 
-from ._requester import RequestExecutor
-from .base_client import BaseClient, Tracer
+from .http_client_base import HTTPClientBase
 
 logger = logging.getLogger(__name__)
 
 
-class Client(BaseClient):
+class Client(HTTPClientBase):
     """
     Core HTTP client for the iMednet API.
 
@@ -35,46 +33,8 @@ class Client(BaseClient):
         backoff_factor: Multiplier for exponential backoff.
     """
 
-    def __init__(
-        self,
-        api_key: Optional[str] = None,
-        security_key: Optional[str] = None,
-        base_url: Optional[str] = None,
-        timeout: Union[float, httpx.Timeout] = 30.0,
-        retries: int = 3,
-        backoff_factor: float = 1.0,
-        log_level: Union[int, str] = logging.INFO,
-        tracer: Optional[Tracer] = None,
-    ) -> None:
-        super().__init__(
-            api_key=api_key,
-            security_key=security_key,
-            base_url=base_url,
-            timeout=timeout,
-            retries=retries,
-            backoff_factor=backoff_factor,
-            log_level=log_level,
-            tracer=tracer,
-        )
-        self._executor = RequestExecutor(
-            lambda *a, **kw: self._client.request(*a, **kw),
-            is_async=False,
-            retries=self.retries,
-            backoff_factor=self.backoff_factor,
-            tracer=self._tracer,
-        )
-
-    def _create_client(self, api_key: str, security_key: str) -> httpx.Client:
-        return httpx.Client(
-            base_url=self.base_url,
-            headers={
-                "Accept": "application/json",
-                "Content-Type": "application/json",
-                "x-api-key": api_key,
-                "x-imn-security-key": security_key,
-            },
-            timeout=self.timeout,
-        )
+    HTTPX_CLIENT_CLS = httpx.Client
+    IS_ASYNC = False
 
     def __enter__(self) -> Client:
         return self
@@ -91,14 +51,6 @@ class Client(BaseClient):
         """Close the underlying HTTP client."""
         self._client.close()
 
-    @property
-    def _should_retry(self) -> Callable[[RetryCallState], bool]:
-        return self._executor.should_retry or self._executor._default_should_retry
-
-    @_should_retry.setter
-    def _should_retry(self, func: Callable[[RetryCallState], bool]) -> None:
-        self._executor.should_retry = func
-
     def get(
         self,
         path: str,
@@ -112,7 +64,7 @@ class Client(BaseClient):
             path: URL path or full URL.
             params: Query parameters.
         """
-        return cast(httpx.Response, self._executor("GET", path, params=params, **kwargs))
+        return cast(httpx.Response, self._request("GET", path, params=params, **kwargs))
 
     def post(
         self,
@@ -127,4 +79,4 @@ class Client(BaseClient):
             path: URL path or full URL.
             json: JSON body for the request.
         """
-        return cast(httpx.Response, self._executor("POST", path, json=json, **kwargs))
+        return cast(httpx.Response, self._request("POST", path, json=json, **kwargs))
